fix: posts table foreign key with user. update: databaseed

+++ database/migrations/2021_05_04_204630_create_posts_table.php
@@ -19,8 +19,8 @@ class CreatePostsTable extends Migration
             $table->longText('description')->nullable();
             $table->timestamps();
 
-            //relation with user table
-            $table->integer('user_id')->unsigned();
+            // //relation with user table
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
